05_django/10-02-django-rest-framework/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.15.0
jedi==0.19.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
